client/server: use server-sent stun servers
Changes the protocol and adds an 'iceServers' which gives the client
a set of STUN/TURN servers to use. The server side implementation is
synchronous and only sending stun servers. If the TURN server credentials
are generated using the shared-secret approach described in
  https://tools.ietf.org/html/draft-uberti-behave-turn-rest-00
it remains a synchronous operation but when using an external provider
credentials might have to be fetched in an async operation.

The client considers the websocket connection only to be completed once
it receives this message. This avoids a race condition where a call is initiated
while the iceServers are not configured yet.

Reminder: there are no free TURN servers, see
  https://bloggeek.me/google-free-turn-server/