change bandwidth on the fly
Shows how to change bandwidth on the fly using the RTCRtpSender.setParameters API.
This is based on
  https://webrtc.github.io/samples/src/content/peerconnection/bandwidth/
and demonstrates how to check for support of certain features as well
as the impact of the bandwidth on the video quality. Also we can see our getStats
logging from the earlier lesson in action.