add stream metadata mapping and signal it
adds a custom metadata mapping to the signalling protocol.
This allows attaching application data to the stream, such as
what kind of content is being shared (for two cameras).

This must use the stream id since the track id is not guaranteed
to be the same on both sides of the connection, see
  https://github.com/w3c/webrtc-pc/issues/1128