turn off the webcam light
with a 5 second delay after hitting the mute button in
order to deal with users who hit the wrong button.

Note that if you get a new peer and send with a video track
in the `ended` readyState, the peer will not receive any
video frames. When using a video element to play audio and
video this may cause audio playout to be blocked. The solution
is to split the element into a video element for the video playout
and a separate audio element for the audio.